Rewrite the app registration
The reimplementation makes the terminal integration cross-platform

+++ Athena/Commands/IntegrationCommands.cs
@@ -11,40 +11,46 @@ public class IntegrationCommands : ICommands
 {
     private readonly AppRunner _runner;
     private readonly string _appPath;
-    private readonly string _symlinkPath = Path.Combine(
-        Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".local", "bin", "athena");
+    private readonly string _appPathDir;
+    private readonly string _symlinkPath;
     
     public IntegrationCommands(AppRunner runner)
     {
         _runner = runner;
         
         // On Linux, the assembly's location points to its dll instead of the executable
-        _appPath = Path.ChangeExtension(Assembly.GetExecutingAssembly().Location, null);
+        var assemblyLocation = Assembly.GetExecutingAssembly().Location;
+        _appPath = Path.ChangeExtension(assemblyLocation,
+            RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? "exe" : null);
+        
+        _symlinkPath = Path.Combine(
+            Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".local", "bin", "athena");
+        
+        _appPathDir = Path.GetDirectoryName(_appPath)!;
     }
     
     [Command("status", Description = "Show the current status of Athena")]
     public void Status()
     {
+        var registrationStatus = "[red]\u25cf Not registered (run `athena integration --add` to register)[/]";
+        
         if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
         {
-            AnsiConsole.MarkupLine("[red]\u25cf This feature isn't supported on Windows![/]");
-            return;
+            var path = Environment.GetEnvironmentVariable("PATH", EnvironmentVariableTarget.User);
+            var pattern = path!.EndsWith(';') ? $"{_appPathDir};" : $";{_appPathDir}";
+            
+            if (path.Contains(pattern))
+                registrationStatus = $"[green]\u25cf Registered[/] at {_appPath}";
         }
-
-        string registrationStatus;
         
-        if (File.Exists(_symlinkPath))
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux) && File.Exists(_symlinkPath))
         {
             var symlinkTarget = File.ResolveLinkTarget(_symlinkPath, true)!.FullName;
             registrationStatus = symlinkTarget == _appPath
-                ? $"[green]\u25cf Registered[/] at {_symlinkPath} --> [green]{_appPath}[/]"
+                ? $"[green]\u25cf Registered[/] at {_symlinkPath}"
                 : $"[darkorange]\u25cf Registered[/] at {_symlinkPath}" +
                   $"\n\tRunning instance: [green]{_appPath}[/]" +
                   $"\n\tSymlink target: [darkorange]{symlinkTarget}[/]";
-        }
-        else
-        {
-            registrationStatus = "[red]\u25cf Not registered (run `athena integration --add` to register)[/]";
         }
         
         AnsiConsole.MarkupLine($"{registrationStatus}");
@@ -55,22 +61,36 @@ public class IntegrationCommands : ICommands
         [Option('a', Description = "Add the integration")] bool add,
         [Option('r', Description = "Remove the integration")] bool remove)
     {
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-        {
-            AnsiConsole.MarkupLine("[red]\u25cf This feature isn't supported on Windows![/]");
-            return;
-        }
-        
         if (add)
         {
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                var pathBefore = Environment.GetEnvironmentVariable("PATH", EnvironmentVariableTarget.User);
+                var pathAfter = pathBefore!.EndsWith(';')
+                    ? $"{pathBefore}{_appPathDir};"
+                    : $"{pathBefore};{_appPathDir}";
+                Environment.SetEnvironmentVariable("PATH", pathAfter, EnvironmentVariableTarget.User);
+                return;
+            }
+            
             if (File.Exists(_symlinkPath)) File.Delete(_symlinkPath);
             await _runner.RunAsync("ln", $"-s {_appPath} {_symlinkPath}");
             return;
         }
 
-        if (remove && File.Exists(_symlinkPath))
+        if (remove)
         {
-            File.Delete(_symlinkPath);
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                var pathBefore = Environment.GetEnvironmentVariable("PATH", EnvironmentVariableTarget.User);
+                var pathAfter = pathBefore!.EndsWith(';')
+                    ? pathBefore.Replace($"{_appPathDir};", "")
+                    : pathBefore.Replace($";{_appPathDir}", "");
+                Environment.SetEnvironmentVariable("PATH", pathAfter, EnvironmentVariableTarget.User);
+                return;
+            }
+            
+            if (File.Exists(_symlinkPath)) File.Delete(_symlinkPath);
             return;
         }
